Update for modern OS
